requirements.txt
botasaurus=4.0.74
python-dotenv=1.0.1
curl_cffi=0.13.0